update: added an AutoUpdateableAdapter interface
