+++ app/src/main/java/com/chydee/lacasadepapel/ui/game/GameFragment.kt
@@ -1,4 +1,4 @@
-package com.chydee.lacasadepapel.game
+package com.chydee.lacasadepapel.ui.game
 
 import android.animation.Animator
 import android.content.Context
@@ -16,7 +16,7 @@ import androidx.navigation.NavOptions
 import androidx.navigation.fragment.findNavController
 import com.chydee.lacasadepapel.R
 import com.chydee.lacasadepapel.databinding.GameFragmentBinding
-import com.chydee.lacasadepapel.models.Quiz
+import com.chydee.lacasadepapel.ui.models.Quiz
 import com.google.android.material.button.MaterialButton
 import kotlinx.android.parcel.RawValue
 import java.util.*
