+++ app/src/main/java/com/chydee/lacasadepapel/ui/models/Quiz.kt
@@ -1,4 +1,4 @@
-package com.chydee.lacasadepapel.models
+package com.chydee.lacasadepapel.ui.models
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
