+++ app/src/main/java/com/chydee/lacasadepapel/utils/AutoUpdatableAdapter.kt
@@ -1,0 +1,26 @@
+package com.chydee.lacasadepapel.utils
+
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.RecyclerView
+
+interface AutoUpdatableAdapter {
+    fun <T> RecyclerView.Adapter<*>.autoNotify(
+        oldList: List<T>,
+        newList: List<T>,
+        compare: (T, T) -> Boolean
+    ) {
+        val diff = DiffUtil.calculateDiff(object : DiffUtil.Callback() {
+            override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
+                return compare(oldList[oldItemPosition], newList[newItemPosition])
+            }
+
+            override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean {
+                return oldList[oldItemPosition] == newList[newItemPosition]
+            }
+
+            override fun getOldListSize() = oldList.size
+            override fun getNewListSize() = newList.size
+        })
+        diff.dispatchUpdatesTo(this)
+    }
+}
