+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.MainActivity">
+        tools:context=".ui.activity.MainActivity">
 
         <fragment
             android:id="@+id/on_boarding_nav_host"
