+++ app/src/main/res/layout/game_result_fragment.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".game.GameResult">
+        tools:context=".ui.game.GameResult">
 
         <TextView
             android:id="@+id/textView"
