+++ app/src/main/res/layout/on_board_fragment.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="onBoardingViewModel"
-            type="com.chydee.lacasadepapel.welcome.OnBoardViewModel" />
+            type="com.chydee.lacasadepapel.ui.OnBoardViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -15,7 +15,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/secondaryColor"
-        tools:context=".welcome.OnBoardFragment">
+        tools:context=".ui.OnBoardFragment">
 
         <TextView
             android:id="@+id/intro_text_textview"
