+++ app/src/main/res/layout/welcome_user_fragment.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="welcomeViewModel"
-            type="com.chydee.lacasadepapel.welcome.WelcomeUserViewModel" />
+            type="com.chydee.lacasadepapel.ui.WelcomeUserViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -15,7 +15,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         style="@style/WelcomeTextStyle"
-        tools:context=".welcome.WelcomeUserFragment">
+        tools:context=".ui.WelcomeUserFragment">
 
         <TextView
             android:id="@+id/welcome_msg"
